Moved tag buttons to own component
